Refactored out EventMessage
This greatly simplifies the consumer, processor, and recorder. So much
so that now the processor is seriously YAGNI.

Before, the consumer pulled messages from rabbit and passed them to the
processor. The processor was another GenServer that any additional
processing we might need could take place. It was parsing the json to
exctract the type out of the message, and so it kind of made sense to
leave it.

Now the message is wrapped up in a EventMessage struct that also
contains some of the rabbit components like the channel, tags, etc. This
has made it trivially easy to ack/nack/reject the messages after
recording them. Also this is the ideal place for the one thing that the
processor was doing.