Take AMQP out of critical path
Fixes #1

Before this commit, if RabbitMQ was not running, the app would not
start. If RabbitMQ was running, the Consumer would get a reference to
the channel, but would not be linked to the connection or channel
process. If the connection was lost, the channel would die, and never
notify the Consumer. Since the Consumer is just waiting for messages, it
would wait forever and never receive anything, even though a new
connection may have been made.

This commit checks the response of creating the connection and loops
after sleeping for 1 second if the connection is not available. It
continues looping until the connection does become available. This
allows the app to start event if RabbitMQ is not running.

This commit also gets a handle to the connection when the consumer is
started, and linkes the consumer process to the connection. If the
connection process dies, then the consumer process is killed and
restarted.