Re-enqueue unack'd messages, add dead letter ex
Fixes #5

Before this commit, if a worker died before writing to elasticsearch,
the message would never be acknowledged, at least not until the
connection was reset.

This adds a `terminate` function to the worker that rejects the message
if the worker dies. If the message hasn't been redelivered previously,
then it is just requeued in the same queue. If it has been redelivered,
then maybe there is an issue with the message itself causing the worker
to die. In that case we send it to the dead letter exchange and let
something else pick it up so a real person can see what happened with
that message.

If the message has already been ack'd and then the worker dies, this
will still try to reject the message. Normally, this would be an issue,
but in elixir/erlang we just let it fail. The worker is already dying,
and any exception that is thrown due to trying to reject a previously
ack'd message is not a problem.