Adds update-able list of Elasticsearch servers
Fixes #3

This adds a `HostManager` GenServer that holds lists of servers. The
list is initially populated from the config, but can be updated while
the app is running.

Currently only used by the archive worker, but should support amqp
without much effort.

The archive worker gets an initial list of elasticsearch servers to
write to. After each request to elasticsearch, the worker checks the
response and if it was successful, it just keeps on with it's current
list. If there was an issue writing to elasticsearch, the message is
requeued, and the archive worker forgets that elasticsearch host. The
next request it will try the next elasticsearch host in it's list. If
the list is empty, only then will it request a new list from the
HostManager.