Update amqp package in prep for exrm
amqp was missing some files needed to build a release with exrm. This
updates the version of amqp to a exrm compatiable version.

Along with the version bump, there were some api changes in the amqp
package. This also addes the necessary changes to keep the code working
with the new version of amqp.